refactor: create useNowPlaying custom hook
